requirements.txt
Django==5.0.4
Pillow==12.1.0
python-dateutil==2.9.0.post0
djangorestframework==3.16.1
django-cors-headers==4.9.0
